backend/requirements/common.txt
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
psycopg2-binary==2.8.4
names==0.3.0